Fix missing image on Twitter Cards
For pages which do not specify an image, we now set a default image so
that Jekyll-seo-tag will set an `og:image` and `twitter:image` tag.

Also set the default card type to "summary" for non-posts in the root
folder (mainly, the index, papers, and blog list pages). This prevents a
REALLY LARGE PICTURE OF MY HEAD from showing up on Twitter posts linking
to those pages.