Make Book Backlinks more selective
Books with simple titles sometimes had false positives if the word was
used in the text. This should make it so we only match on real links.

There is one edge case where it will match if you inlcude a
book_link.html but don't use the variable. I can't figure out how to get
around that. Yet.