Update author and series generation
I now actually parse the YAML frontmatter, instead of cheating and
looking for keywords. Keyword search fails on a couple of pages that
have the word "series:" outside of the frontmatter.

Also change the series and author pages to automatically generate their
intro paragraph from their frontmatter, instead of hard coding the
paragraph in the script.

This fixed #74.