Auto-insert superscript commas between adjacent footnotes
Previously we did this with CSS, but it failed if footnotes were on
different words in the same paragraph, the second one would have a
comman when it shouldn't.

Instead we now use Liquid `replace` in the layout to find adjacent
footnote references (`</a></sup><sup `) and inserts a superscripted
comma (`<sup class="fn-separator">, </sup>`) between them.

This fixes #79.