Add inverted favorites map to link cache
Enhance the `LinkCacheGenerator` to build a second, inverted map for
"favorites" posts.

In addition to the `favorites_mentions` map (`book_url => [post]`), the
generator now creates a `favorites_posts_to_books` map (`post_url =>
[book]`).

This new map is optimized for the upcoming "By Award" page enhancement,
allowing for efficient lookup of all books featured in a given
"favorites" post. Tests have been added to verify the structure and
correctness of this new cache.